Add search headlines
